Map GFW data for a geartype, above a minimum level of fishing activity (polar projection)

This function saves a static map of fishing activity, with a CRS appropriate for the arctic.

Usage

jacks_polar_static(gear, limit)

Arguments

gear

The gear type used by boats to be mapped.

limit

The minimum number of hours of fishing activity required for a pixel to be included in the map.

Value

The function saves a map of fishing activity for a particular gear type above a minimum level of activity.

Details

The function saves a map of fishing activity, as recorded by global fishing watch, with FAO fishing regions overlayed. The user can specify which fishing gear type to visualise, and can also remove points below a certain amount of fishing effort. This is useful for keeping the plot clean, but also allows you to remove any fishing signals you might not trust from the global fishing watch model GFW model.

See also